requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
Django==3.2
djangorestframework==3.14.0
filelock==3.8.2
idna==3.4
mysqlclient==2.1.1
platformdirs==2.6.0
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.17.1
